Comparison operators of FilingSet are not defined for bad iterables.
Comparison operators are only defined if all the values in the compared iterable are Filing objects. They are not defined for non-iterables. '==' and '!=' always return a bool though.

Checking if items in iterable are Filing haves the values as a list and thus iterates only once. This means that generators can also be used with operators and other 'normal' class methods.