Added initial Sphinx documentation and workflow for GH Pages.
In pyproject.toml, removed ruff dependency as it is nowadays shipped with hatch. Removed trove classifiers for frameworks (this library is not an extension of AsyncIO, Hatch, Pytest etc.). Updated documentation URL. Added doc environment and created scripts 'build' and 'fbuild' (Windows only, removes generated folders and build first).

Sphinx documentation uses theme furo and it is defined as a dependency for the doc environment and HTML build script ammaraskar/sphinx-action as requirements.txt file in 'docs/source' in workflow. Release defined as '1.0a1'.